% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpp2019-package.R
\docType{package}
\name{kpp2019-package}
\alias{kpp2019}
\alias{kpp2019-package}
\title{kpp2019: Kenya Population Projections 2019}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides data from the Kenya National Bureau of Statistics Projections 2019.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://kpp2019.damurka.com}
  \item \url{https://github.com/damurka/kpp2019}
  \item \url{https://damurka.github.io/kpp2019/}
  \item \url{http://kpp2019.damurka.com/}
  \item Report bugs at \url{https://github.com/damurka/kpp2019/issues}
}

}
\author{
\strong{Maintainer}: David Kariuki \email{hello@damurka.com} (\href{https://orcid.org/0009-0003-6159-1107}{ORCID}) [copyright holder]

}
\keyword{internal}
